requirements.txt
customtkinter==5.1.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
pycountry==22.3.5
pygame==2.3.0
scikit_learn==1.2.2
seaborn==0.12.2